PLU JSON
This is the JSON conversion of the same data.

https://csvjson.com/csv2json

When demoing this concept eariler today, I tried out a few npm libraries to do it for me, but they didn't work and or work as expected, so I went back to the classic way of doing things, and found a web app tool to do it instead. Reminds me of my early programming days :)

Turns out there are duplicate entries for a two PLU values, 4014 and 4388. Going to see about which I should keep in this dataset, or if it is safe to merge the metadata for the two.