Valencia Orange Dedupe
Going with what appears to be essentially the same entry, for the most part. I want to have single entries for each PLU, for the most part. Unless this comes up again and it is detrimental in keeping a bunch of extra metadata for a lot of items, then I will roll back on this one, and keep both somehow.