Produce Type
Now you can use a type to figure out what kind of produce you have, based on the category, and what kinds of specifiers it has. Say like it's in the Nuts category, it will only possibly be a few different kinds of products. This is a very similar setup to that of what I described for Minecraft's NBT type definitions. It does appear to get fairly slow for TS's type checking unfortunately, I think I could probably workout a way to declare things a bit better.